requirements.txt
json
python-docx
tqdm
gradio
csv
textwrap
BeautifulSoup4
markdown
pandas
librosa
google-generativeai
pathlib
pydub